fix build on Windows
fixes the following error from CI:

common\QskAspect.cpp(199): error C2666: 'QFlags<QskAspect::State>::operator &': 3 overloads have similar conversions
D:\a\qskinny\Qt\5.15.2\msvc2019_64\include\QtCore/qflags.h(146): note: could be 'QFlags<QskAspect::State> QFlags<QskAspect::State>::operator &(Enum) noexcept const'
        with
        [
            Enum=QskAspect::State
        ]
D:\a\qskinny\Qt\5.15.2\msvc2019_64\include\QtCore/qflags.h(145): note: or       'QFlags<QskAspect::State> QFlags<QskAspect::State>::operator &(uint) noexcept const'
D:\a\qskinny\Qt\5.15.2\msvc2019_64\include\QtCore/qflags.h(144): note: or       'QFlags<QskAspect::State> QFlags<QskAspect::State>::operator &(int) noexcept const'
common\QskAspect.cpp(199): note: or       'built-in C++ operator&(QFlags<QskAspect::State>::Int, quint16)'
common\QskAspect.cpp(199): note: while trying to match the argument list '(QskAspect::States, const quint16)'

+++ src/common/QskAspect.cpp
@@ -194,7 +194,7 @@ static QByteArray qskStatesToString(
 
     for ( int i = 0; i < 16; i++ )
     {
-        const quint16 mask = 1 << i;
+        const uint mask = 1 << i;
 
         if ( states & mask )
         {
